feat: 공고 마감 Usecase 추가
1. closeJobPosting() 추가
2. JobPostingUseCaseImpl 파일 분리
